Ung - Henter data for satser og ungomdsprogram først når panelet åpnes Preloader data slik at den er klar dersom aksjonspunkt først Forhindrer tregere lasting dersom aksjonspunkt
